feat(client): Add serialized response format option to Presto client (#26654)
Summary:

The capability exists for Presto server to return serialized responses, but is currently not enabled in the client nor piped through to the CLI. As far as I know, there is no OSS client which can read using the serialized interface. It makes it kind of annoying to test out any changes in this area

Adding support in the client to request and process serialized responses from the server

To support the new client processing, I created a TypeManager and binary deserializer for use in the client

NOTE: right now I only handle the types located in presto-common on the serialized path; there are some additional types declared in presto-main-base which the client does not have access to. Those could either be moved or redefined in presto-client, but since serialized APIs are opt-in and fails fast, it seemed acceptable to defer support for those types

Differential Revision: D87268210